Update requirements document
The original author provided a txt file for installing the necessary
dependencies. This file is very old and has not been maintained for
a long time. It was not possible to use this file and follow the steps
provided by the author to successfully install the dependencies on
our machine. This may be due to the fact that some of the older
versions of the libraries mentioned in the file are no longer supported
or are not compatible with the current platform we are using.

Find requirements that are valid for the current platform and document
them in a requirements file to replace old requirements that are no
longer valid. Replace the txt format originally used with a yml format
that is more conda compliant and better suited for cross-platform
installation.